# src/main/resources/openApi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Sample Project with OpenApi",
        "description": "Sample Project having mongo db and openapi",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "paths": {
        "/api/student/update": {
            "put": {
                "tags": [
                    "student-controller"
                ],
                "operationId": "updateDetails",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Student"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Student"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/student/create": {
            "post": {
                "tags": [
                    "student-controller"
                ],
                "operationId": "createStudent",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Student"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Student"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/student/startsWith": {
            "get": {
                "tags": [
                    "student-controller"
                ],
                "operationId": "getByNameStartsWIth",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Student"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/student/partial": {
            "get": {
                "tags": [
                    "student-controller"
                ],
                "operationId": "getPartialStudents",
                "parameters": [
                    {
                        "name": "pageNo",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "pageLimit",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Student"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/student/getByNameEmail": {
            "get": {
                "tags": [
                    "student-controller"
                ],
                "operationId": "getStudentByNameAndEmail",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Student"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/student/getByName/{name}": {
            "get": {
                "tags": [
                    "student-controller"
                ],
                "operationId": "getStudentByName",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Student"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/student/get/{id}": {
            "get": {
                "tags": [
                    "student-controller"
                ],
                "operationId": "logInfo",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Student"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/student/emailLike": {
            "get": {
                "tags": [
                    "student-controller"
                ],
                "operationId": "getByEmailLike",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Student"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/student/bySubName": {
            "get": {
                "tags": [
                    "student-controller"
                ],
                "operationId": "getBySubjectName",
                "parameters": [
                    {
                        "name": "subName",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Student"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/student/byDeptName": {
            "get": {
                "tags": [
                    "student-controller"
                ],
                "operationId": "getByDeptName",
                "parameters": [
                    {
                        "name": "deptName",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Student"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/student/all": {
            "get": {
                "tags": [
                    "student-controller"
                ],
                "operationId": "getAllStudent",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Student"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/student/allSorted": {
            "get": {
                "tags": [
                    "student-controller"
                ],
                "operationId": "getSortedResult",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Student"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/student/delete/{id}": {
            "delete": {
                "tags": [
                    "student-controller"
                ],
                "operationId": "deleteStudent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Department": {
                "type": "object",
                "properties": {
                    "deptName": {
                        "type": "string"
                    },
                    "location": {
                        "type": "string"
                    }
                }
            },
            "Student": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "department": {
                        "$ref": "#/components/schemas/Department"
                    },
                    "subjects": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Subject"
                        }
                    }
                }
            },
            "Subject": {
                "type": "object",
                "properties": {
                    "subName": {
                        "type": "string"
                    },
                    "maxMarks": {
                        "type": "number",
                        "format": "double"
                    },
                    "compulsory": {
                        "type": "boolean"
                    }
                }
            }
        }
    }
}